refactor: restructure Home module
Home now has two subpackages: api & impl. Code is currently in impl.
