+++ feature/home/impl/src/main/kotlin/com/potatosheep/kite/feature/home/impl/nav/HomeNavigation.kt
@@ -1,9 +1,9 @@
-package com.potatosheep.kite.feature.library.nav
+package com.potatosheep.kite.feature.home.impl.nav
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.potatosheep.kite.core.common.enums.SortOption
-import com.potatosheep.kite.feature.library.HomeRoute
+import com.potatosheep.kite.feature.home.impl.HomeRoute
 
 @Composable
 fun HomeScreen(
