+++ feature/home/impl/src/main/res/mipmap-anydpi/ic_launcher.xml
