+++ feature/home/impl/src/main/res/mipmap-mdpi/ic_launcher.webp
