+++ settings.gradle.kts
@@ -33,7 +33,8 @@ include(":feature:feed:api")
 include(":core:ui")
 include(":core:designsystem")
 include(":feature:subscription")
-include(":feature:home")
+include(":feature:home:impl")
+include(":feature:home:api")
 include(":feature:post")
 include(":feature:image")
 include(":feature:subreddit")
